Make the extension usable in any language.
Because of the code search in the bookmark tree explicit for the children called "other bookmarks", the extension is not applicable for user with another language. Through these little changes it should be possible to use it in any language. One you can note is that the folder may will be in another order, but at least now it would be possible to rename the folder.